Fixed typos, removed unused code and refactored for readability

+++ src/avr/Lin.cpp
@@ -65,7 +65,7 @@ uint8_t Lin::request(uint8_t address, uint8_t *data, uint8_t length)
     delay(length + 1U);
     uint8_t bytesReceived{0U};
     int16_t byte{0U};
-    const uint8_t idByte{(address & 0x3FU) | addressParity(address)};
+    const uint8_t idByte{static_cast<uint8_t>((address & 0x3FU) | addressParity(address))};
     int16_t countdown{static_cast<int16_t>(124'000'000UL / baud)};
     serialBreak();
     Serial.write(0x55U);
